+++ src/network/validator_credentials.go
@@ -40,20 +40,11 @@ func (c *ValidatorCredentials) Print() {
 	utils.ColoredPrintln("Validator Index To:", c.ValidatorIndexTo)
 }
 
-func (c *ValidatorCredentials) CreateNodeRecovery() NodeRecovery {
-	return NodeRecovery{
-		ValidatorMnemonic:  c.ValidatorMnemonic,
-		WithdrawalMnemonic: c.WithdrawalMnemonic,
-		KeystoreIndexFrom:  c.ValidatorIndexFrom,
-		KeystoreIndexTo:    c.ValidatorIndexTo,
-	}
-}
-
 func (c *ValidatorCredentials) FromNodeRecovery(nr NodeRecovery) *ValidatorCredentials {
-	c.ValidatorIndexTo = nr.KeystoreIndexTo
-	c.ValidatorIndexFrom = nr.KeystoreIndexFrom
-	c.WithdrawalMnemonic = nr.WithdrawalMnemonic
-	c.ValidatorMnemonic = nr.ValidatorMnemonic
+	c.ValidatorIndexTo = nr.ValidatorCredentials.ValidatorIndexTo
+	c.ValidatorIndexFrom = nr.ValidatorCredentials.ValidatorIndexFrom
+	c.WithdrawalMnemonic = nr.ValidatorCredentials.WithdrawalMnemonic
+	c.ValidatorMnemonic = nr.ValidatorCredentials.ValidatorMnemonic
 	return c
 }
 
@@ -73,30 +64,63 @@ func (c *ValidatorCredentials) GenerateMnemonic() error {
 	if err != nil {
 		return err
 	}
-
+	useExisting, err := UseExistingMnemonicPrompt()
+	if err != nil {
+		return err
+	}
 	fmt.Println("Generating mnemonic")
-
-	output, err := GetMnemonic()
+	output, err := GetMnemonic(useExisting)
 	if err != nil {
 		return err
 	}
 	c.ValidatorMnemonic = output
 	c.WithdrawalMnemonic = output
 
-	propmt := promptui.Select{
+	err = c.GenerateWithdrawalCredentials()
+	if err != nil {
+		return err
+	}
+
+	fmt.Println("A mnemonic was generated and stored in node_config.yaml.\n Make sure you don't loose it as you will not be able to recover your keystore if you loose it....")
+	return nil
+}
+
+func (c *ValidatorCredentials) GenerateWithdrawalCredentials() error {
+	promptWithdrawal := promptui.Select{
 		Label: "Generate separate withdrawal mnemonic? [Yes/No]",
 		Items: []string{"Yes", "No"},
 	}
-	_, generateVal, err := propmt.Run()
+	_, generateVal, err := promptWithdrawal.Run()
 	if err != nil {
 		return err
 	}
 	if generateVal == "Yes" {
-		c.WithdrawalMnemonic, err = GetMnemonic()
+		c.WithdrawalMnemonic, err = GetMnemonic(false)
 		if err != nil {
 			return err
 		}
 	}
+	return nil
+}
+
+func (c *ValidatorCredentials) GenerateMnemonicWithoutPrompt() error {
+	err := CheckAndDownloadValTool()
+	if err != nil {
+		return err
+	}
+
+	fmt.Println("Generating mnemonic")
+
+	output, err := GetMnemonic(false)
+	if err != nil {
+		return err
+	}
+	c.ValidatorMnemonic = output
+	output, err = GetMnemonic(false)
+	if err != nil {
+		return err
+	}
+	c.WithdrawalMnemonic = output
 
 	fmt.Println("A mnemonic was generated and stored in node_config.yaml.\n Make sure you don't loose it as you will not be able to recover your keystore if you loose it....")
 	return nil